+++ srcs/lookAt.c
@@ -6,7 +6,7 @@
 /*   By: tamigore <tamigore@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/08 10:41:56 by tamigore          #+#    #+#             */
-/*   Updated: 2022/04/14 16:06:09 by tamigore         ###   ########.fr       */
+/*   Updated: 2022/04/14 17:12:54 by tamigore         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,13 +20,13 @@
 
 static t_vec		vector_direct(int fov, int resX, int resY, int x, int y)
 {
-	double		Px;
-	double		Py;
-	double		ratio;
-	double		tang;
+	float		Px;
+	float		Py;
+	float		ratio;
+	float		tang;
 
 	tang = tan((fov * M_PI / 180) / 2);
-	ratio = (double)resX / (double)resY;
+	ratio = (float)resX / (float)resY;
 	Px = (2 * ((x + 0.5) / resX) - 1) * ratio * tang;
 	Py = (1 - 2 * ((y + 0.5) / resY)) * tang;
 	return (vec_init(Px, Py, -1, 0));
